+++ capacitor.config.ts
@@ -10,11 +10,15 @@ const config: CapacitorConfig = {
   plugins: {
     SplashScreen: {
       androidSplashResourceName: "splash"
-    }
+    },
+    Keyboard: {
+      style: "DARK",
+      resizeOnFullScreen: true,
+    },
   },
   // Use this for live reload https://capacitorjs.com/docs/guides/live-reload
   // server: {
-  //   url: "http://192.168.1.250:1847",
+  //   url: "http://192.168.1.115:1847",
   //   cleartext: true
   // },
 };