+++ src/app/components/PeopleItem.svelte
@@ -1,16 +1,23 @@
 <script lang="ts">
   import {onMount} from "svelte"
-  import {ctx} from "@welshman/lib"
+  import {formatTimestampRelative} from "@welshman/lib"
   import type {Filter} from "@welshman/util"
   import {deriveEvents} from "@welshman/store"
-  import {repository, load, loadRelaySelections, formatTimestampRelative} from "@welshman/app"
+  import {load} from "@welshman/net"
+  import {Router} from "@welshman/router"
+  import {repository, loadRelaySelections} from "@welshman/app"
   import Icon from "@lib/components/Icon.svelte"
   import Link from "@lib/components/Link.svelte"
   import Profile from "@app/components/Profile.svelte"
   import ProfileInfo from "@app/components/ProfileInfo.svelte"
-  import {makeChatPath} from "@app/routes"
+  import {pubkeyLink} from "@app/state"
 
-  const {pubkey} = $props()
+  type Props = {
+    pubkey: string
+    url?: string
+  }
+
+  const {pubkey, url}: Props = $props()
 
   const filters: Filter[] = [{authors: [pubkey], limit: 1}]
   const events = deriveEvents(repository, {filters})
@@ -22,27 +29,27 @@
     // Load at least one note, regardless of time frame
     load({
       filters: [{authors: [pubkey], limit: 1}],
-      relays: ctx.app.router.FromPubkeys([pubkey]).getUrls(),
+      relays: Router.get().FromPubkeys([pubkey]).getUrls(),
     })
   })
 </script>
 
 <div class="card2 bg-alt col-2 shadow-xl">
   <div class="flex justify-between">
-    <Profile {pubkey} />
-    <Link class="btn btn-primary hidden sm:flex" href={makeChatPath([pubkey])}>
-      <Icon icon="letter" />
-      Start a Chat
+    <Profile {pubkey} {url} />
+    <Link external href={pubkeyLink(pubkey)} class="btn btn-primary hidden sm:flex">
+      <Icon icon="user-circle" />
+      See Complete Profile
     </Link>
   </div>
-  <ProfileInfo {pubkey} />
+  <ProfileInfo {pubkey} {url} />
   {#if $events.length > 0}
     <div class="bg-alt badge badge-neutral border-none">
       Last active {formatTimestampRelative($events[0].created_at)}
     </div>
   {/if}
-  <Link class="btn btn-primary sm:hidden" href={makeChatPath([pubkey])}>
-    <Icon icon="letter" />
-    Start a Chat
+  <Link external href={pubkeyLink(pubkey)} class="btn btn-primary sm:hidden">
+    <Icon icon="user-circle" />
+    See Complete Profile
   </Link>
 </div>