+++ src/app/editor/index.ts
@@ -1,27 +1,65 @@
-import "@welshman/editor/index.css"
-
 import {mount} from "svelte"
 import type {Writable} from "svelte/store"
 import {get} from "svelte/store"
-import {Editor} from "@tiptap/core"
-import {ctx} from "@welshman/lib"
 import type {StampedEvent} from "@welshman/util"
-import {signer, profileSearch} from "@welshman/app"
-import {MentionSuggestion, WelshmanExtension} from "@welshman/editor"
-import {getSetting, userSettingValues} from "@app/state"
-import {MentionNodeView} from "./MentionNodeView"
+import {makeEvent, getTagValues, getListTags, BLOSSOM_AUTH} from "@welshman/util"
+import {simpleCache, normalizeUrl, removeNil, now} from "@welshman/lib"
+import {Router} from "@welshman/router"
+import {signer, profileSearch, userBlossomServers} from "@welshman/app"
+import {Editor, MentionSuggestion, WelshmanExtension} from "@welshman/editor"
+import {makeMentionNodeView} from "./MentionNodeView"
 import ProfileSuggestion from "./ProfileSuggestion.svelte"
 import {PermalinkExtension} from "nostr-git"
 import Spinner from "@src/lib/components/Spinner.svelte"
 
-export const getUploadType = () => getSetting<"nip96" | "blossom">("upload_type")
+export const hasBlossomSupport = simpleCache(async ([url]: [string]) => {
+  const $signer = signer.get()
+  const headers: Record<string, string> = {
+    "X-Content-Type": "text/plain",
+    "X-Content-Length": "1",
+    "X-SHA-256": "73cb3858a687a8494ca3323053016282f3dad39d42cf62ca4e79dda2aac7d9ac",
+  }
 
-export const getUploadUrl = () => {
-  const {upload_type, nip96_urls, blossom_urls} = userSettingValues.get()
+  try {
+    if ($signer) {
+      const event = await signer.get().sign(
+        makeEvent(BLOSSOM_AUTH, {
+          tags: [
+            ["t", "upload"],
+            ["server", url],
+            ["expiration", String(now() + 30)],
+          ],
+        }),
+      )
 
-  return upload_type === "nip96"
-    ? nip96_urls[0] || "https://nostr.build"
-    : blossom_urls[0] || "https://cdn.satellite.earth"
+      headers.Authorization = `Nostr ${btoa(JSON.stringify(event))}`
+    }
+
+    const res = await fetch(normalizeUrl(url) + "/upload", {method: "head", headers})
+
+    return res.status === 200
+  } catch (e) {
+    if (!String(e).includes("Failed to fetch")) {
+      console.error(e)
+    }
+  }
+
+  return false
+})
+
+export const getUploadUrl = async (spaceUrl?: string) => {
+  const userUrls = getTagValues("server", getListTags(userBlossomServers.get()))
+  const allUrls = removeNil([spaceUrl, ...userUrls])
+
+  for (let url of allUrls) {
+    url = url.replace(/^ws/, "http")
+
+    if (await hasBlossomSupport(url)) {
+      return url
+    }
+  }
+
+  return "https://cdn.satellite.earth"
 }
 
 export const signWithAssert = async (template: StampedEvent) => {
@@ -30,26 +68,30 @@ export const signWithAssert = async (template: StampedEvent) => {
   return event!
 }
 
-export const makeEditor = ({
+export const makeEditor = async ({
   aggressive = false,
   autofocus = false,
   charCount,
   content = "",
   placeholder = "",
+  url,
   submit,
   uploading,
   wordCount,
+  disableFileUpload,
 }: {
   aggressive?: boolean
   autofocus?: boolean
   charCount?: Writable<number>
   content?: string
   placeholder?: string
+  url?: string
   submit: () => void
   uploading?: Writable<boolean>
   wordCount?: Writable<number>
-}) =>
-  new Editor({
+  disableFileUpload?: boolean
+}) => {
+  return new Editor({
     content,
     autofocus,
     element: document.createElement("div"),
@@ -62,8 +104,8 @@ export const makeEditor = ({
       WelshmanExtension.configure({
         submit,
         sign: signWithAssert,
-        defaultUploadType: getUploadType(),
-        defaultUploadUrl: getUploadUrl(),
+        defaultUploadType: "blossom",
+        defaultUploadUrl: await getUploadUrl(url),
         extensions: {
           placeholder: {
             config: {
@@ -75,29 +117,31 @@ export const makeEditor = ({
               aggressive,
             },
           },
-          fileUpload: {
-            config: {
-              onDrop() {
-                uploading?.set(true)
+          fileUpload: disableFileUpload
+            ? false
+            : {
+                config: {
+                  onDrop() {
+                    uploading?.set(true)
+                  },
+                  onComplete() {
+                    uploading?.set(false)
+                  },
+                },
               },
-              onComplete() {
-                uploading?.set(false)
-              },
-            },
-          },
           nprofile: {
             extend: {
-              addNodeView: () => MentionNodeView,
+              addNodeView: () => makeMentionNodeView(url),
               addProseMirrorPlugins() {
                 return [
                   MentionSuggestion({
                     editor: (this as any).editor,
                     search: (term: string) => get(profileSearch).searchValues(term),
-                    getRelays: (pubkey: string) => ctx.app.router.FromPubkeys([pubkey]).getUrls(),
+                    getRelays: (pubkey: string) => Router.get().FromPubkeys([pubkey]).getUrls(),
                     createSuggestion: (value: string) => {
                       const target = document.createElement("div")
 
-                      mount(ProfileSuggestion, {target, props: {value}})
+                      mount(ProfileSuggestion, {target, props: {value, url}})
 
                       return target
                     },
@@ -114,3 +158,4 @@ export const makeEditor = ({
       charCount?.set(editor.storage.wordCount.chars)
     },
   })
+}