+++ src/lib/components/ContentSearch.svelte
@@ -8,15 +8,13 @@
   const {...props}: Props = $props()
 </script>
 
-<div class="col-2 content-padding-t content-padding-x h-full {props.class}">
+<div class="content-padding-t content-padding-x flex h-full flex-col gap-1 {props.class}">
   <div class="z-feature">
     <div class="content-sizing">
       {@render props.input?.()}
     </div>
   </div>
-  <div class="scroll-container overflow-auto pt-2">
-    <div class="content-sizing">
-      {@render props.content?.()}
-    </div>
+  <div class="scroll-container content-sizing overflow-auto pt-2">
+    {@render props.content?.()}
   </div>
 </div>