+++ src/lib/components/PageBar.svelte
@@ -1,18 +1,20 @@
 <script lang="ts">
+  import type {Snippet} from "svelte"
+
   interface Props {
-    icon?: import("svelte").Snippet
-    title?: import("svelte").Snippet
-    action?: import("svelte").Snippet
+    icon?: Snippet
+    title?: Snippet
+    action?: Snippet
     [key: string]: any
   }
 
   const {...props}: Props = $props()
 </script>
 
-<div class="relative z-feature mx-2 rounded-xl pt-4 {props.class}">
+<div data-component="PageBar" class="cw top-sai fixed z-feature p-2">
   <div
-    class="flex min-h-12 items-center justify-between gap-4 rounded-xl bg-base-100 px-4 shadow-xl">
-    <div class="flex items-center gap-4">
+    class="flex min-h-12 items-center justify-between gap-4 rounded-xl rounded-xl bg-base-100 px-4 shadow-xl">
+    <div class="ellipsize flex items-center gap-4 whitespace-nowrap">
       {@render props.icon?.()}
       {@render props.title?.()}
     </div>