+++ src/lib/util.ts
@@ -3,20 +3,6 @@ import * as nip19 from "nostr-tools/nip19"
 import {range, DAY} from "@welshman/lib"
 import { getEventTags } from "@welshman/util"
 
-export const displayList = <T>(xs: T[], conj = "and", n = 6, locale = "en-US") => {
-  const stringItems = xs.map(String)
-
-  if (xs.length > n + 2) {
-    const formattedList = new Intl.ListFormat(locale, {style: "long", type: "unit"}).format(
-      stringItems.slice(0, n),
-    )
-
-    return `${formattedList}, ${conj} ${xs.length - n} others`
-  }
-
-  return new Intl.ListFormat(locale, {style: "long", type: "conjunction"}).format(stringItems)
-}
-
 export const nsecEncode = (secret: string) => nip19.nsecEncode(hexToBytes(secret))
 
 export const nsecDecode = (nsec: string) => {
@@ -30,7 +16,6 @@ export const nsecDecode = (nsec: string) => {
 export const day = (seconds: number) => Math.floor(seconds / DAY)
 
 export const daysBetween = (start: number, end: number) => [...range(start, end, DAY)].map(day)
-
 
 export const FREELANCE_JOB = 32767
   // GIT_REPOSITORY is mistakenly defined in welshman as 30403 which
@@ -55,3 +40,5 @@ export const getRootEventTagValue = (tags: string[][]): string | undefined => {
     t => t[3]==='root' || t[2] ==='root'
   )?.[1]
 }
+
+export const ucFirst = (s: string) => s.slice(0, 1).toUpperCase() + s.slice(1)